requirements.txt
appdirs==1.4.3
attrs==18.2.0
black==18.9b0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Django==2.1.5
django-filter==2.0.0
djangorestframework==3.9.1
httpie==1.0.2
idna==2.8
Pygments==2.3.1
pytz==2018.7
requests==2.21.0
toml==0.10.0
urllib3==1.24.1
django-rest-multiple-models == 2.1.3
